+++ src/react/forms.jsx
@@ -8,8 +8,7 @@
 var util = require('../js/lib/util.js');
 
 var textfieldClass = 'mui-textfield',
-    textfieldInputClass = 'mui-textfield__input',
-    floatingLabelClass = 'mui-textfield__label--floating',
+    floatMod = '--float-label',
     isEmptyClass = 'mui--is-empty',
     isNotEmptyClass = 'mui--is-not-empty',
     isDirtyClass = 'mui--is-dirty';
@@ -24,6 +23,7 @@ var Input = React.createClass({
     return {
       value: '',
       type: 'text',
+      placeholder: '',
       autofocus: false,
       onChange: null
     };
@@ -39,7 +39,6 @@ var Input = React.createClass({
         isNotEmpty = Boolean(this.state.value),
         inputEl;
 
-    cls[textfieldInputClass] = true;
     cls[isEmptyClass] = !isNotEmpty;
     cls[isNotEmptyClass] = isNotEmpty;
     cls[isDirtyClass] = this.state.isDirty;
@@ -51,6 +50,7 @@ var Input = React.createClass({
         <textarea
           ref="input"
           className={ cls }
+          placeholder={ this.props.placeholder }
           value={ this.props.value }
           autoFocus={ this.props.autofocus }
           onChange={ this._handleChange }
@@ -64,6 +64,7 @@ var Input = React.createClass({
           className={ cls }
           type={ this.props.type }
           value={ this.state.value }
+          placeholder={ this.props.placeholder }
           autoFocus={ this.props.autofocus }
           onChange={ this._handleChange }
           onFocus={ this._handleFocus }
@@ -122,7 +123,6 @@ var Label = React.createClass({
     return (
       <label
         refs="label"
-        className={ (this.props.floating) ? floatingLabelClass : '' }
         style={ this.state.style }
         onClick={ this.props.onClick }
       >
@@ -145,29 +145,35 @@ var Textfield = React.createClass({
       type: 'text',
       value: '',
       label: '',
+      placeholder: '',
       isLabelFloating: false,
       autofocus: false,
       onChange: null
     };
   },
   render: function() {
-    var labelEl;
+    var cls = {},
+        labelEl;
 
     if (this.props.label.length) {
       labelEl = (
         <Label
           text={ this.props.label }
           onClick={ this._focus }
-          floating={ this.props.isLabelFloating }
         />
       );
     }
 
+    cls[textfieldClass] = true;
+    cls[textfieldClass + floatMod] = this.props.isLabelFloating;
+    cls = util.classNames(cls);
+
     return (
-      <div className={ textfieldClass }>
+      <div className={ cls }>
         <Input
           type={ this.props.type }
           value={ this.props.value }
+          placeholder={ this.props.placeholder }
           autoFocus={ this.props.autofocus }
           onChange={ this.props.onChange }
         />